pas_archive/pyproject.toml
[tool.poetry]
name = "pas-bot"
version = "0.1.0"
description = "Therapeutic bot for supporting alienated parents"
authors = ["PAS Bot Team"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.7"
langchain = "^0.1.0"
langgraph = "^0.0.20"
langchain-openai = "^0.0.5"
nemo-guardrails = "^0.7.0"
guardrails-ai = "^0.3.0"
transformers = "^4.36.0"
torch = "^2.0.0"
detoxify = "^0.5.1"
redis = "^5.0.0"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
presidio-analyzer = "^2.2.0"
presidio-anonymizer = "^2.2.0"
natasha = "^1.6.0"
spacy = "^3.7.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
structlog = "^24.1.0"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
httpx = "^0.26.0"
apscheduler = "^3.10.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^24.0.0"
ruff = "^0.1.0"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true